Rename a few things.
