+++ .gitignore
@@ -15,6 +15,8 @@
 *~
 /package_build_*
 /release
+MANUAL
+MANUAL.html
 README
 README.html
 /unifi-poller_manual.html
