+++ cmd/unifi-poller/main.go
@@ -1,29 +1,14 @@
 package main
 
 import (
-	"fmt"
 	"log"
-	"os"
 
-	unifipoller "github.com/davidnewhall/unifi-poller/pkg/unifi-poller"
+	"github.com/davidnewhall/unifi-poller/unifipoller"
 )
 
+// Keep it simple.
 func main() {
-	log.SetFlags(log.LstdFlags)
-	if err := run(); err != nil {
+	if err := unifipoller.Start(); err != nil {
 		log.Fatalln("[ERROR]", err)
 	}
-}
-
-func run() error {
-	unifi := &unifipoller.UnifiPoller{}
-	if unifi.ParseFlags(os.Args[1:]); unifi.ShowVer {
-		fmt.Printf("unifi-poller v%s\n", unifipoller.Version)
-		return nil // don't run anything else w/ version request.
-	}
-	if err := unifi.GetConfig(); err != nil {
-		unifi.Flag.Usage()
-		return err
-	}
-	return unifi.Run()
 }
