+++ examples/MANUAL.md
